[🌱 feat] 사용자 질문 분류 로직 구현 - 사용자 질문을 상품 추천 / 서비스 문의 / 일반 지식 3가지로 분류하는 프롬프트 작성 - 분류 결과에 따라 챗봇 응답 분기 처리 로직 추가
