+++ src/main/java/org/beep/sbpp/chatbot/service/ChatbotEmbeddingService.java
@@ -1,0 +1,12 @@
+package org.beep.sbpp.chatbot.service;
+
+import org.beep.sbpp.chatbot.dto.ProductVectorDTO;
+
+import java.util.List;
+
+public interface ChatbotEmbeddingService {
+
+    void vectorizeInit();
+    void addProduct(ProductVectorDTO dto);
+    void addProducts(List<ProductVectorDTO> list);
+}
