--- src/main/java/org/beep/sbpp/chatbot/service/EmbeddingInitService.java
@@ -1,4 +1,0 @@
-package org.beep.sbpp.chatbot.service;
-
-public interface EmbeddingInitService {
-}
