--- src/main/java/org/beep/sbpp/chatbot/service/EmbeddingInitServiceImpl.java
@@ -1,4 +1,0 @@
-package org.beep.sbpp.chatbot.service;
-
-public class EmbeddingInitServiceImpl implements EmbeddingInitService {
-}
